See README.md for details.

+++ immortals_repo/shared/modules/dfus/ElevationApi-1/src/main/java/mil/darpa/immortals/dfus/ElevationData.java
@@ -17,16 +17,16 @@ public class ElevationData {
         return hae;
     }
 
-    public long getLe() {
+    public double getLe() {
         return le;
     }
 
     private final double latitude;
     private final double longitude;
     private final double hae;
-    private final long le;
+    private final double le;
 
-    public ElevationData(double latitude, double longitude, double hae, long le) {
+    public ElevationData(double latitude, double longitude, double hae, double le) {
         this.latitude = latitude;
         this.longitude = longitude;
         this.hae = hae;
